refactor(theme): migrate from contexts to feature-based architecture
