+++ src/features/auth/index.ts
@@ -1,0 +1,11 @@
+// Auth feature barrel export
+// TODO: Implement auth components, hooks, and services
+
+// Placeholder exports for future implementation
+// export { LoginForm } from './components/LoginForm'
+// export { SignupForm } from './components/SignupForm'
+// export { useAuth } from './hooks/useAuth'
+// export { authSlice } from './store/authSlice'
+
+// Types
+// export type { User, AuthState } from './types'
