+++ src/features/workout/index.ts
@@ -1,0 +1,11 @@
+// Workout feature barrel export
+// TODO: Implement workout components, hooks, and services
+
+// Placeholder exports for future implementation
+// export { WorkoutForm } from './components/WorkoutForm'
+// export { ExerciseList } from './components/ExerciseList'
+// export { useWorkout } from './hooks/useWorkout'
+// export { workoutSlice } from './store/workoutSlice'
+
+// Types
+// export type { Workout, Exercise, WorkoutSet } from './types'
