+++ src/shared/components/layout/index.ts
@@ -1,0 +1,3 @@
+// Layout components barrel export
+export { AppHeader } from './AppHeader'
+export { AppSidebar } from './AppSidebar'
