+++ src/shared/utils/index.ts
@@ -1,0 +1,1 @@
+//  Utilitaires
